Use dedicated cache for error code range parsing
Some code has been reshuffled around to avoid circular imports
